refactor(admin): consolidate settings and api-key module
General settings, API-key management, the settings composable and styles were split across generic folders. Move Settings.vue, ApiKeys.vue, useSettings.ts and settings.scss under admin/settings so configuration screens are one module.
