+++ webui/src/admin/settings/settings.scss
